move GAdsPrefetchTask out of SystemObserver
