+++ Branch-SDK-TestBed/src/main/AndroidManifest.xml
@@ -64,11 +64,11 @@
         </activity>
 
         <meta-data android:name="io.branch.sdk.TestMode" android:value="true" /> <!-- Set to true to use Branch_Test_Key -->
-<!--        <meta-data android:name="io.branch.sdk.BranchKey" android:value="key_live_feebgAAhbH9Tv85H5wLQhpdaefiZv5Dv" />-->
-<!--        <meta-data android:name="io.branch.sdk.BranchKey.test" android:value="key_test_hdcBLUy1xZ1JD0tKg7qrLcgirFmPPVJc" />-->
+        <meta-data android:name="io.branch.sdk.BranchKey" android:value="key_live_feebgAAhbH9Tv85H5wLQhpdaefiZv5Dv" />
+        <meta-data android:name="io.branch.sdk.BranchKey.test" android:value="key_test_hdcBLUy1xZ1JD0tKg7qrLcgirFmPPVJc" />
         <!--    BMF keys, to test limited access products like CPID that are only enabled on BMF    -->
-        <meta-data android:name="io.branch.sdk.BranchKey" android:value="key_live_hkDytPACtipny3N9XmnbZlapBDdj4WIL" />
-        <meta-data android:name="io.branch.sdk.BranchKey.test" android:value="key_test_jpAAsHxzDjint6I30mkBZjeewAki21R7" />
+<!--        <meta-data android:name="io.branch.sdk.BranchKey" android:value="key_live_hkDytPACtipny3N9XmnbZlapBDdj4WIL" />-->
+<!--        <meta-data android:name="io.branch.sdk.BranchKey.test" android:value="key_test_jpAAsHxzDjint6I30mkBZjeewAki21R7" />-->
 
         <!-- Optional. Set to true to disable auto deep link feature-->
         <meta-data android:name="io.branch.sdk.auto_link_disable" android:value="false"/>
